Refactor: make section in sidenavigation (#320)
