+++ src/components/layout/SideNavigation/SideNavigation.tsx
@@ -3,10 +3,9 @@
 import { useSideNavStore } from '@/stores/sideNavStore';
 import { motion } from 'framer-motion';
 
-import SideNavigationBottom from './components/Bottom/SideNavigationBottom';
-import ChatRoomSection from './components/ChatRoomSection/ChatRoomSection';
 import SideNavigationHeader from './components/Header/SideNavigationHeader';
 import MenuSection from './components/MenuSection/MenuSection';
+import SideNavigationBottom from './components/SideNavBottom/SideNavBottom';
 
 export default function SideNavigation() {
   const { isOpen, toggle } = useSideNavStore();
@@ -18,11 +17,10 @@ export default function SideNavigation() {
         transition={{ type: 'spring', stiffness: 200, damping: 24 }}
         className="bg-gray50 flex h-screen flex-col justify-between overflow-hidden p-5 shadow-md"
       >
-        <div className="flex h-full flex-col overflow-hidden">
-          <SideNavigationHeader isOpen={isOpen} onClick={toggle} />
-          <MenuSection isOpen={isOpen} />
-          <div className="custom-scrollbar flex min-h-0 flex-1 flex-col overflow-y-auto">
-            {isOpen && <ChatRoomSection />}
+        <div className="flex h-full flex-col justify-between">
+          <div>
+            <SideNavigationHeader isOpen={isOpen} onClick={toggle} />
+            <MenuSection isOpen={isOpen} />
           </div>
           <SideNavigationBottom />
         </div>
